+++ lib/src/utils/get_suitable_calendar_view_grid_view_builder_widget.dart
@@ -3,13 +3,15 @@ import 'package:clean_calendar/src/ui/calendar_view.dart';
 import 'package:flutter/material.dart';
 
 class GetSuitableCalendarViewGridViewBuilderWidget extends StatelessWidget {
-  const GetSuitableCalendarViewGridViewBuilderWidget(
-      {Key? key,
-      required this.pageViewDate,
-      required this.datePickerCalendarView})
-      : super(key: key);
+  const GetSuitableCalendarViewGridViewBuilderWidget({
+    Key? key,
+    required this.pageViewDate,
+    required this.startWeekday,
+    required this.datePickerCalendarView,
+  }) : super(key: key);
 
   final DateTime pageViewDate;
+  final int startWeekday;
   final DatePickerCalendarView datePickerCalendarView;
 
   @override
@@ -17,11 +19,13 @@ class GetSuitableCalendarViewGridViewBuilderWidget extends StatelessWidget {
     if (datePickerCalendarView == DatePickerCalendarView.monthView) {
       return MonthViewCalendarDatesSection(
         pageViewDate: pageViewDate,
+        startWeekday: startWeekday,
       );
     }
     if (datePickerCalendarView == DatePickerCalendarView.weekView) {
       return WeekViewCalendarDatesSection(
         pageViewDate: pageViewDate,
+        startWeekday: startWeekday,
       );
     }
     // if (dateSelectionMode == DatePickerSelectionMode.range) {